Final realization of the course project

+++ src/main/java/by/bsuir/acc_reg_system/logic/comand/DeleteOrderCommand.java
@@ -12,9 +12,6 @@ import java.sql.SQLException;
 import java.util.Collection;
 import java.util.Iterator;
 
-/**
- * Created by Vladislav on 12.05.15.
- */
 public class DeleteOrderCommand implements Command {
 
 
@@ -23,47 +20,58 @@ public class DeleteOrderCommand implements Command {
 
         String idString = req.getParameter("idDelete");
 
-        int id = Integer.parseInt(idString);
+        if (idString.equals("")) {
 
-        Orders order = Factory.getInstance().getOrderDAO().getOrderById(id);
-
-        if (order.getTemplate() != null) {
-
-            Collection templates = Factory.getInstance().getTemplateDAO().getTemplatesByOrder(order);
-
-            Iterator iterator1 = templates.iterator();
-
-            while (iterator1.hasNext()) {
-
-                Template template = (Template) iterator1.next();
-
-                Collection products = Factory.getInstance().getProductDAO().getProductsByTemplate(template);
-
-                Iterator iterator2 = products.iterator();
-
-                while (iterator2.hasNext()) {
-
-                    Product product = (Product) iterator2.next();
-                    Factory.getInstance().getProductDAO().deleteProduct(product);
-                }
-
-                Factory.getInstance().getTemplateDAO().deleteTemplate(template);
-            }
+            req.getSession(true).setAttribute("idDelete", "Enter a number!");
 
         } else {
 
-            Collection ownerTemplates = Factory.getInstance().getOwnerTemplateDAO().getOwnerTemplatesByOrder(order);
+            int id = Integer.parseInt(idString);
 
-            Iterator iterator3 = ownerTemplates.iterator();
+            Orders order = Factory.getInstance().getOrderDAO().getOrderById(id);
 
-            while (iterator3.hasNext()) {
+            if (order != null) {
 
-                OwnerTemplate ownerTemplate = (OwnerTemplate) iterator3.next();
 
-                Factory.getInstance().getOwnerTemplateDAO().deleteOwnerTemplate(ownerTemplate);
+                if (order.getTemplate() != null) {
+
+                    Collection templates = Factory.getInstance().getTemplateDAO().getTemplatesByOrder(order);
+
+                    Iterator iterator1 = templates.iterator();
+
+                    while (iterator1.hasNext()) {
+
+                        Template template = (Template) iterator1.next();
+
+                        Collection products = Factory.getInstance().getProductDAO().getProductsByTemplate(template);
+
+                        Iterator iterator2 = products.iterator();
+
+                        while (iterator2.hasNext()) {
+
+                            Product product = (Product) iterator2.next();
+                            Factory.getInstance().getProductDAO().deleteProduct(product);
+                        }
+
+                        Factory.getInstance().getTemplateDAO().deleteTemplate(template);
+                    }
+
+                } else {
+
+                    Collection ownerTemplates = Factory.getInstance().getOwnerTemplateDAO().getOwnerTemplatesByOrder(order);
+
+                    Iterator iterator3 = ownerTemplates.iterator();
+
+                    while (iterator3.hasNext()) {
+
+                        OwnerTemplate ownerTemplate = (OwnerTemplate) iterator3.next();
+
+                        Factory.getInstance().getOwnerTemplateDAO().deleteOwnerTemplate(ownerTemplate);
+                    }
+                }
+
+                Factory.getInstance().getOrderDAO().deleteOrder(order);
             }
         }
-
-        Factory.getInstance().getOrderDAO().deleteOrder(order);
     }
 }
